fix(mouse_handler): apply a couple of fixes to mouse wheel event
First of all, this updates the interactor with the mouse position on a wheel event so that if there
are multiple renderers, the interactor can figure out which one needs to be updated.

Second, this forwards the event to the interactor, rather than applying a manual zoom to the
camera ourselves. This makes the behavior more consistent.

Third, this skips the zoom for the start event, since there appears to always be a wheel event
right after it.

Fixes: pyvista/pyvista#4020

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>