chore(js):  Allow triggering release on external artifact static_viewer.html
Previously, when a new version of static_viewer.html is released on the master
of vtk-js we couldn't trigger a release because we do not commit this file but
rather grab it during release. This commit adds the hash of the latest version
in a file and a hash check that runs during release but does not trigger a
failure on a mismatch but rather a warning.
So now if static_viewer.html is updated on vtk-js:
- If we do not care `.fetch_externals.sh` will just emit a warning
- If we **do** care we update .static_viewer.sha256 and trigger a new release
  on trame-vtk even if no other `trame-vtk` change happened during last
  release.